refactor(resolver): cns api
