+++ apps/wallet-mobile/src/features/Send/useCases/StartMultiTokenTx/InputReceiver/ShowResolvedAddressSelected.stories.tsx
@@ -8,6 +8,7 @@ import {Boundary} from '../../../../../components'
 import {SelectedWalletProvider} from '../../../../../SelectedWallet'
 import {YoroiWallet} from '../../../../../yoroi-wallets/cardano/types'
 import {mocks as walletMocks} from '../../../../../yoroi-wallets/mocks/wallet'
+import {CardanoMobile} from '../../../../../yoroi-wallets/wallets'
 import {initialState, SendProvider, SendState} from '../../../common/SendContext'
 import {ShowResolvedAddressSelected} from './ShowResolvedAddressSelected'
 
@@ -32,6 +33,7 @@ const Wrapper = ({ns}: {ns: Partial<SendState>}) => {
         apiKey: 'apiKey',
       },
     },
+    csl: CardanoMobile,
   })
   const resolverStorage = resolverStorageMaker()
   const resolverManager = resolverManagerMaker(resolverStorage, resolverApi)
