+++ apps/wallet-mobile/src/features/Send/useCases/StartMultiTokenTx/SelectNameServer/SelectNameServer.stories.tsx
@@ -8,6 +8,7 @@ import {Boundary} from '../../../../../components'
 import {SelectedWalletProvider} from '../../../../../SelectedWallet'
 import {YoroiWallet} from '../../../../../yoroi-wallets/cardano/types'
 import {mocks as walletMocks} from '../../../../../yoroi-wallets/mocks/wallet'
+import {CardanoMobile} from '../../../../../yoroi-wallets/wallets'
 import {initialState, SendProvider, SendState} from '../../../common/SendContext'
 import {SelectNameServer} from './SelectNameServer'
 
@@ -31,6 +32,7 @@ const UnselectedNS = () => {
         apiKey: 'apiKey',
       },
     },
+    csl: CardanoMobile,
   })
   const resolverStorage = resolverStorageMaker()
   const resolverManager = resolverManagerMaker(resolverStorage, resolverApi)
@@ -53,6 +55,7 @@ const SelectedNS = () => {
         apiKey: 'apiKey',
       },
     },
+    csl: CardanoMobile,
   })
   const resolverStorage = resolverStorageMaker()
   const resolverManager = resolverManagerMaker(resolverStorage, resolverApi)
