+++ apps/wallet-mobile/translations/messages/src/TxHistory/TxHistoryNavigator.json
@@ -4,14 +4,14 @@
     "defaultMessage": "!!!Receive",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 348,
+      "line": 350,
       "column": 16,
-      "index": 13066
+      "index": 13147
     },
     "end": {
-      "line": 351,
+      "line": 353,
       "column": 3,
-      "index": 13155
+      "index": 13236
     }
   },
   {
@@ -19,14 +19,14 @@
     "defaultMessage": "!!!Swap",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 352,
+      "line": 354,
       "column": 13,
-      "index": 13170
+      "index": 13251
     },
     "end": {
-      "line": 355,
+      "line": 357,
       "column": 3,
-      "index": 13243
+      "index": 13324
     }
   },
   {
@@ -34,14 +34,14 @@
     "defaultMessage": "!!!Swap from",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 356,
+      "line": 358,
       "column": 17,
-      "index": 13262
+      "index": 13343
     },
     "end": {
-      "line": 359,
+      "line": 361,
       "column": 3,
-      "index": 13339
+      "index": 13420
     }
   },
   {
@@ -49,14 +49,14 @@
     "defaultMessage": "!!!Swap to",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 360,
+      "line": 362,
       "column": 15,
-      "index": 13356
+      "index": 13437
     },
     "end": {
-      "line": 363,
+      "line": 365,
       "column": 3,
-      "index": 13429
+      "index": 13510
     }
   },
   {
@@ -64,14 +64,14 @@
     "defaultMessage": "!!!Slippage Tolerance",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 364,
+      "line": 366,
       "column": 21,
-      "index": 13452
+      "index": 13533
     },
     "end": {
-      "line": 367,
+      "line": 369,
       "column": 3,
-      "index": 13547
+      "index": 13628
     }
   },
   {
@@ -79,14 +79,14 @@
     "defaultMessage": "!!!Select pool",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 368,
+      "line": 370,
       "column": 14,
-      "index": 13563
+      "index": 13644
     },
     "end": {
-      "line": 371,
+      "line": 373,
       "column": 3,
-      "index": 13644
+      "index": 13725
     }
   },
   {
@@ -94,14 +94,14 @@
     "defaultMessage": "!!!Send",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 372,
+      "line": 374,
       "column": 13,
-      "index": 13659
+      "index": 13740
     },
     "end": {
-      "line": 375,
+      "line": 377,
       "column": 3,
-      "index": 13739
+      "index": 13820
     }
   },
   {
@@ -109,14 +109,14 @@
     "defaultMessage": "!!!Scan QR code address",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 376,
+      "line": 378,
       "column": 18,
-      "index": 13759
+      "index": 13840
     },
     "end": {
-      "line": 379,
+      "line": 381,
       "column": 3,
-      "index": 13860
+      "index": 13941
     }
   },
   {
@@ -124,14 +124,14 @@
     "defaultMessage": "!!!Select asset",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 380,
+      "line": 382,
       "column": 20,
-      "index": 13882
+      "index": 13963
     },
     "end": {
-      "line": 383,
+      "line": 385,
       "column": 3,
-      "index": 13971
+      "index": 14052
     }
   },
   {
@@ -139,14 +139,14 @@
     "defaultMessage": "!!!Selected tokens",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 384,
+      "line": 386,
       "column": 26,
-      "index": 13999
+      "index": 14080
     },
     "end": {
-      "line": 387,
+      "line": 389,
       "column": 3,
-      "index": 14103
+      "index": 14184
     }
   },
   {
@@ -154,14 +154,14 @@
     "defaultMessage": "!!!Edit amount",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 388,
+      "line": 390,
       "column": 19,
-      "index": 14124
+      "index": 14205
     },
     "end": {
-      "line": 391,
+      "line": 393,
       "column": 3,
-      "index": 14217
+      "index": 14298
     }
   },
   {
@@ -169,14 +169,14 @@
     "defaultMessage": "!!!Confirm",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 392,
+      "line": 394,
       "column": 16,
-      "index": 14235
+      "index": 14316
     },
     "end": {
-      "line": 395,
+      "line": 397,
       "column": 3,
-      "index": 14321
+      "index": 14402
     }
   },
   {
@@ -184,14 +184,14 @@
     "defaultMessage": "!!!Share this address to receive payments. To protect your privacy, new addresses are generated automatically once you use them.",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 396,
+      "line": 398,
       "column": 19,
-      "index": 14342
+      "index": 14423
     },
     "end": {
-      "line": 402,
+      "line": 404,
       "column": 3,
-      "index": 14580
+      "index": 14661
     }
   },
   {
@@ -199,14 +199,14 @@
     "defaultMessage": "!!!Confirm transaction",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 403,
+      "line": 405,
       "column": 27,
-      "index": 14609
+      "index": 14690
     },
     "end": {
-      "line": 406,
+      "line": 408,
       "column": 3,
-      "index": 14702
+      "index": 14783
     }
   },
   {
@@ -214,14 +214,14 @@
     "defaultMessage": "!!!Please scan a QR code",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 407,
+      "line": 409,
       "column": 13,
-      "index": 14717
+      "index": 14798
     },
     "end": {
-      "line": 410,
+      "line": 412,
       "column": 3,
-      "index": 14792
+      "index": 14873
     }
   },
   {
@@ -229,14 +229,14 @@
     "defaultMessage": "!!!Success",
     "file": "src/TxHistory/TxHistoryNavigator.tsx",
     "start": {
-      "line": 411,
+      "line": 413,
       "column": 25,
-      "index": 14819
+      "index": 14900
     },
     "end": {
-      "line": 414,
+      "line": 416,
       "column": 3,
-      "index": 14893
+      "index": 14974
     }
   }
 ]
