+++ packages/resolver/src/adapters/api.test.ts
@@ -1,4 +1,5 @@
 import {Resolver} from '@yoroi/types'
+import {init} from '@emurgo/cross-csl-nodejs'
 
 import {resolverApiMaker} from './api'
 
@@ -8,6 +9,7 @@ const mockApiConfig = {
       apiKey: 'mock-api-key',
     },
   },
+  csl: init('ctx'),
 }
 const mockError = new Error('Test Error')
