--- packages/resolver/src/adapters/cns/types/cnsMetadata.ts
@@ -1,11 +1,0 @@
-export interface CNSMetadata {
-  name: string
-  image: string
-  expiry: number
-  origin: string
-  cnsType: string
-  mediaType: string
-  description: string
-  virtualSubdomainLimits: number
-  virtualSubdomainEnabled: 'Enabled' | 'Disabled'
-}
