--- packages/resolver/src/adapters/cns/types/index.ts
@@ -1,4 +1,0 @@
-export * from './cnsMetadata'
-export * from './cnsUserRecord'
-export * from './plutus'
-export * from './socialRecords'
