--- packages/resolver/src/adapters/cns/types/socialRecords.ts
@@ -1,9 +1,0 @@
-export type SocialRecord =
-  | 'NICKNAME'
-  | 'EMAIL'
-  | 'TELEGRAM'
-  | 'PHONE_NUMBER'
-  | 'DISCORD'
-  | 'GITHUB'
-  | 'MEDIUM'
-  | 'TWITTER'
